Install fix

+++ ops_scripts/orchestra/install.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 keyfile="~/.ssh/wei-dkfz.pem"
 
+if [ ! -f ~/.orchestra_subnet ]; then
+    echo "Create your orchestra subnet file."
+    exit 1
+fi
+
 subnet=`cat ~/.orchestra_subnet`
 echo "Installing dependencies ..."
-echo "" > ~/.orchestra_subnet
 sudo apt-get install -y python-pip 2>&1 > /dev/null
 sudo pip install netaddr 2>&1 > /dev/null
-sudo cp orchestra.py /bin/orchestra
-sudo chmod +x /bin/orchestra
+sudo sudo ln -s orchestra.py /bin/orchestra
+sudo chmod +x orchestra.py
 
 python subnet-install.py $subnet $keyfile
